Collapse collectors
As DF (original) and BF (added in 1.8) started as copy-paste of each other, they
share a LOT. This PR tries to minimize redundancy, still, not loosing clarity,
so leaving inner Args class for each of them.
